+++ src/DbTraffic.Core/Enums/ObjectAccessType.cs
@@ -1,0 +1,8 @@
+namespace DbTraffic.Core.Enums;
+
+public enum ObjectAccessType
+{
+    Read = 1,
+    Write = 2,
+    Mixed = 3
+}